# openapi.yaml
openapi: 3.0.0
info:
  title: Indikara Backend API
  version: 1.0.0
  description: API for authentication, products, and health checks
servers:
  - url: https://backend-wei5.onrender.com
    description: Deployment server
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/products:
    get:
      summary: Get all products
      tags:
        - Products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    id:
                      type: number
                    name:
                      type: string
                    price:
                      type: array
                      items:
                        type: object
                        properties:
                          size:
                            type: string
                          amount:
                            type: number
                    image:
                      type: array
                      items:
                        type: string
                    story:
                      type: string
                    description:
                      type: string
                    details:
                      type: array
                      items:
                        type: string
                    color:
                      type: array
                      items:
                        type: string
                    weavingTechnique:
                      type: string
                    material:
                      type: string
                    manufacturer:
                      type: string
                    tags:
                      type: array
                      items:
                        type: string
                    SKU:
                      type: string
                    category:
                      type: string
                    categoryId:
                      type: number
                    createdAt:
                      type: string
                    updatedAt:
                      type: string
        '500':
          description: Server error
    post:
      summary: Create a new product
      tags:
        - Products
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
                - price
                - image
                - description
                - manufacturer
                - SKU
                - category
                - categoryId
              properties:
                id:
                  type: number
                name:
                  type: string
                price:
                  type: array
                  items:
                    type: object
                    properties:
                      size:
                        type: string
                      amount:
                        type: number
                image:
                  type: array
                  items:
                    type: string
                story:
                  type: string
                description:
                  type: string
                details:
                  type: array
                  items:
                    type: string
                color:
                  type: array
                  items:
                    type: string
                weavingTechnique:
                  type: string
                material:
                  type: array
                  items:
                    type: string
                manufacturer:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                SKU:
                  type: string
                category:
                  type: string
                categoryId:
                  type: number
            examples:
              demo:
                value:
                  id: 1
                  name: Demo Product
                  price:
                    - size: M
                      amount: 100
                  image:
                    - https://example.com/image.jpg
                  description: A demo product
                  manufacturer: Demo Maker
                  SKU: DEMO-001
                  category: textiles
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  id:
                    type: number
                  name:
                    type: string
                  price:
                    type: array
                    items:
                      type: object
                      properties:
                        size:
                          type: string
                        amount:
                          type: number
                  image:
                    type: array
                    items:
                      type: string
                  story:
                    type: string
                  description:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
                  color:
                    type: array
                    items:
                      type: string
                  weavingTechnique:
                    type: string
                  material:
                    type: string
                  manufacturer:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  SKU:
                    type: string
                  category:
                    type: string
                  categoryId:
                    type: number
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
        '400':
          description: Invalid product data
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /api/products/{id}:
    get:
      summary: Get product by ID
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  id:
                    type: number
                  name:
                    type: string
                  price:
                    type: array
                    items:
                      type: object
                      properties:
                        size:
                          type: string
                        amount:
                          type: number
                  image:
                    type: array
                    items:
                      type: string
                  story:
                    type: string
                  description:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
                  color:
                    type: array
                    items:
                      type: string
                  weavingTechnique:
                    type: string
                  material:
                    type: string
                  manufacturer:
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
                  SKU:
                    type: string
                  category:
                    type: string
                  categoryId:
                    type: number
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
        '404':
          description: Product not found
        '500':
          description: Server error
...